feat: api.js 분리
